+++ ticketguardian/abstract/tests/test_mixins.py
@@ -65,3 +65,4 @@ def test_list_resource():
 
             for attr in obj:
                 assert hasattr(resource_objects[0], attr)
+